rename ContentHeader as ContentWrapper
