+++ src/components/viewers/CellPanel.tsx
@@ -1,6 +1,6 @@
 import { DataFrame, stringify } from 'hightable'
 import { useEffect, useState } from 'react'
-import ContentHeader from './ContentHeader.js'
+import ContentWrapper from './ContentWrapper.js'
 
 interface ViewerProps {
   df: DataFrame
@@ -54,7 +54,7 @@ export default function CellPanel({ df, row, col, setProgress, setError, onClose
     <span>row {row + 1}</span>
   </>
 
-  return <ContentHeader headers={headers}>
+  return <ContentWrapper headers={headers}>
     <code className="text">{text}</code>
-  </ContentHeader>
+  </ContentWrapper>
 }
