+++ src/components/viewers/index.ts
@@ -1,8 +1,8 @@
-import ContentHeader from './ContentHeader.js'
+import ContentWrapper from './ContentWrapper.js'
 import ImageView from './ImageView.js'
 import MarkdownView from './MarkdownView.js'
 import ParquetView from './ParquetView.js'
 import SlidePanel from './SlidePanel.js'
 import TextView from './TextView.js'
 import Viewer from './Viewer.js'
-export { ContentHeader, ImageView, MarkdownView, ParquetView, SlidePanel, TextView, Viewer }
+export { ContentWrapper, ImageView, MarkdownView, ParquetView, SlidePanel, TextView, Viewer }
